Widen narrow character, affect and object fields to int
The spell, duration, modifier and specific fields of affected_type, the
ability scores, the point fields (armor, spell resistance, hit and
damage roll, saves, resistances), the player's sex, class, level,
race, subrace, weight and height, and the object type, proficiency and
material were char or short while nearly every producer computes an
int. Values past the narrow range were silently truncated; a duration
over 32767 rounds, for example, wrapped negative.

Player and object files are ASCII and read through int temporaries,
no MySQL column stores these fields, and nothing takes their address,
so no format or layout changes. Three duration multipliers and one
height computation now cast explicitly to float.

GCC 16.2 conversion falls from 2463 to 1267 sites and Clang 22.1.8
implicit-int-conversion from 1613 to 527.

+++ src/act.informative.c
@@ -5757,7 +5757,7 @@ static void display_identity_section(struct char_data *ch, int line_length)
 static void display_vitals_section(struct char_data *ch, int line_length)
 {
   struct time_info_data playing_time;
-  float height = GET_HEIGHT(ch);
+  float height = (float)GET_HEIGHT(ch);
 
   skore_section_header(ch, "\tR*** VITALS & CONDITION ***\tC", line_length, "\tC");
